3. Nicer printing
Instead of using 'print' to format the output of listDirectory,
let's try to tidy up the output a bit:

instead of using `print` which figures out how to print everything
automatically - and does an acceptable but ugly job of it - instead
we'll make a nicer list.

The goal for this section is to print one filename per line, without
any quotes around it. To do that we've got two things:

for_, which can iterate over any `Foldable` - which a list is
and putStrLn - which prints a string on a line with no quotes
around it.

So we can put the two together as:

F.for_ files putStrLn

(not as in these commits making a single-point function)

The type of `files` is [FilePath], which is really [String]
and the type of putStrLn is String -> IO ()

So we're going to go over the list, apply putStrLn to each
element which gives us an IO () action, and then run those
IO actions. That's a bit of a complicated way of explaining
a for-loop... the big point is that this for loop isn't
built into the language.

* iterating over lists - there are lots of different ways to learn about (for example, map is the classic functional programming one, and on top of things like that, more complicated patterns are built)

* passing in functions (without applying the parameters)