4. File metadata
Now we're going to show more information about each file,
in addition to each name.

Now, a string isn't going to be enough to hold that information
so we'll make a new data type to hold the information.

part of this -- but just the size bit.
need to add in sorting and _isdir later...

* new datatype
* ugly printing