5. Pretty printing file sizes
File info is ugly in general.

That file size is pretty ugly - it's in bytes.
Let's try to make it a bit more attractive:
if the number is over 1024 bytes, turn this into
kilobytes (or kibibytes, if you prefer)

sequence:
5.1 format with kilobyte(s) and byte(s)
5.2 now get rid of (s) for 1024 and 1 cases and use singular (and no () on the plural)
5.3 factor this into a pluraliser and a unit-izer -- unitizer has to happen first

This is purely functional: we aren't in IO here.

Use view patterns for the pluraliser, viewing the unitizer values.

* introduces pattern guards
* introduces language extensions
* introduces view patterns