7. Adding another option
We can add another option to the parser.

The option we'll add is --long/-l which will allow use to
choose long format output with file sizees, or otherwise
default to short format output with just the names.

But now the option type cannot just be String - we need
a richer type. We could use a tuple, or we can use a named
datatype with named fields.

I choose the latter because it is easier as that structure
gets more complicated.