9.1 Pass the config around inside ReaderT
We've threaded the options around a bit... with this repeated
pattern of always passing 'opt' around in lots of places. if we
had a bigger application, we'd be passing this maybe several
levels down...

maybe we can do something about that.

the repeated pattern I want to eliminate is: functions that
take an option spec and return an IO action.

Use ReaderT for this - our actions were previously
"actions that can do IO" (the specific IO we've used so far
is is reading directories, printing, looking at commandline
options.