12. Add a sort config option
The previous step has various different ways we might want to
sort, but we can only compile with one.

It makes sense to make this configurable at runtime.

We already have optparse-applicative, and we already thread a
config options object around using ReaderT.

So lets use those.

As part of this, we'll turn on stronger type checking to
catch pattern matching bugs that turn into runtime errors.