13. Filter hidden files
On unix, files begining with a dot are usually
hidden by default; but our tool displays them.

This step adds in a configuration option to
filter them out.

There is really nothing much in this step that
we haven't seen before - so maybe if there is
time it is interesting to try this as an
ask-the-audience?